修改外部引用，增强代码易读性

+++ User/Receive.h
@@ -5,6 +5,6 @@ extern int num;
 extern float sum;
 
 char ReceiveVis(void);
-uint8_t ReceiveNum(void);
+float ReceiveNum(void);
 void ReceiveNum_Gettarget(void);
 #endif
